Add classification fields to enrichment if library changes classification data without having classification change.
Previously classification data was only added to enrichment if there was a classification change (i.e. which shelf the book should be on). This meant a library was unable to update a classification field if that change didn't trigger a classification change.

This has been deemed too inconvenient for the libraries, so it has now been changed. The change is that for FBS libraries the full subfields in the classification fields are used for determining if classification data should be included in the enrichment. 
Note that if a classification field is seen as change then all classification fields will be added to the enrichment.

No change in logic regarding automatic creation of enrichments when the common record is updated.

git-svn-id: https://svn.dbc.dk/repos/opencat-business/trunk@118176 b87a6194-a221-0410-ad6f-3b8ec35bba13