Add run-test script using docker
Running jsshell inside a docker makes it possible to run the script on
a machine which doesn't have the jsshell package install - which is more
or less all other machines than devel8.

It would be nice to use this script on Jenkins build but for some
unknown reason the files becomes owned by root which means the next
build will fail as it can't clean up the files. That is a problem for
another time.