Add docker-compose for running systemtest
This is the first version of the systemtest setup.
Currently the test is run by starting the update-deployer image where
the opencat-business/distributions dir is overwritten with the local
dir.
ocb-tools is then used to execute the test as usual.

In this first version the update-payara-deployer is staging but it will
probably be changed to something else once the updateservice build has
been fixed.