Add branch name to compose name
Since multiple tests can run at the same time all the containers need
unique names which means the branch name has to be part of the name.

Also reorganized the runConfigurations so it is possible to run js-test
both with and without docker since the docker method doesn't work on
devel8.