Check for self reference
If a record points to it self in 014 *a it will result in validation
errors, but none of the validation messages explicit says 'you can't
reference the same record' but rather some derived error e.g 'you can't
change from volume to head record'.

This check validates all subfields which have a reference to another
record, which is not already validated, to make sure the record doesn't
reference the value in 001 *a.