Feed getValidateSchemas with library rules
In order to optimized performance of getValidateSchemas the function is not fed with all library rules for the given library. This way it is not longer necessary to call Java for each schema.

Instead of starting with loading every schema we only load those schemas which match the template set.

Auditors: jbn, jbr, mib