Simplify the 014 validation in dbcartanm template
Instead of complicating the validation rule even further it seems to work just fine to simply remove part of the validation.

Reason: If 014 *x contains ANM the record will always point to a 870970 record no matter which record is being validated. If 014 *a has any other (allowed) value then the reference is always to a record with the same agency.
When LookUpRecord doesn't find a match for agency id it defaults to the value of 001 *b of this record, therefore removing the explicit ends up having the intended effect.

Auditors: jbn, jbr, mib, hwha, kpn